moved test scoreboard function to own file
